+++ changelog.md
@@ -2,6 +2,16 @@
 
 All notable changes to `Ezypay` will be documented in this file.
 
+## Version 1.0.4
+
+### Fixed
+- Namespace for WebHookEventTypes enum
+
+## Version 1.0.3
+
+### Added
+- Tests for Webhook security
+
 ## Version 1.0.2
 
 ### Added